requirements.txt
colorama
argon2-cffi>=20.1.0
